Updating detail page

+++ backend/src/main/java/com/rakakusan/shopee_price_history/dto/response/PriceHistoryResponse.java
@@ -1,5 +1,6 @@
 package com.rakakusan.shopee_price_history.dto.response;
 
+import java.math.BigDecimal;
 import java.time.LocalDate;
 
 import lombok.Data;
@@ -8,6 +9,6 @@ import lombok.Data;
 public class PriceHistoryResponse {
     private Long id;
     private Integer price;
-    private Integer discount;
+    private BigDecimal discount;
     private LocalDate recordDate;
 }
